Create PseudoModal model and revamp time input

+++ Caffeinator/Updater.swift
@@ -98,12 +98,10 @@ class Updater {
     /// Displays a dialog prompting the user to update. If the user decides to update, the default browser is launched to download the DMG and the app is terminated to avoid issues when the user attempts to overwrite it; if the update is declined, swap out the update timer for one with a longer time interval (we don't want to be hounding the user to update if they don't want to right now)
     func showUpdatePrompt(fromVersion currentVersion: String, toVersion newVersion: String, withURLString urlString: String, releaseNotes: NSAttributedString) {
         DispatchQueue.main.async {
-            let windowCtrl = (NSApp.delegate as! AppDelegate).storyboard.instantiateController(withIdentifier: "updatePanelController") as? NSWindowController
-            let updateVC = windowCtrl?.contentViewController as! UpdatePanelViewController
-            updateVC.currentVersion = currentVersion
-            updateVC.newVersion = newVersion
-            updateVC.releaseNotes = releaseNotes
-            updateVC.onUpdateConfirmed = {
+            let properties = ["currentVersion": currentVersion,
+                              "newVersion": newVersion,
+                              "releaseNotes": releaseNotes] as [String : Any]
+            let confirmationHandler = { (_: Any?) in
                 if let url = URL(string: urlString) {
                     NSWorkspace.shared.open(url)
                     NSApp.terminate(self)
@@ -111,14 +109,13 @@ class Updater {
                     Notifier.showErrorMessage(withTitle: txt("U.url-open-failure-title"), text: txt("U.url-open-failure-msg"))
                 }
             }
-            updateVC.onUpdatePostponed = {
+            let postponementHandler = {
                 self.updateTimer.invalidate()
                 self.updateTimer = Timer.scheduledTimer(withTimeInterval: self.resumptionDelay, repeats: false) { _ in
                     self.checkForUpdate(isUserInitiated: false)
                 }
             }
-            windowCtrl?.showWindow(self)
-            NSApp.activate(ignoringOtherApps: true)
+            (NSApp.delegate as! AppDelegate).storyboard?.instantiateAndShowPseudoModal(withIDString: "updatePanelController", properties: properties, onConfirm: confirmationHandler, onCancel: postponementHandler)
         }
     }
 }
